week1

+++ Week 1/w1_q1.cpp
@@ -1,9 +1,9 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main(){
-    int n; 
+    long long n; 
     cin>>n;
-  int  sum=n*(n+1)/2;
+   long long  sum= n*(n+1)/2;
   int x;
   while(n>1){
     cin>>x;
