+++ Week 1/w1_q5.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 int main(){
-    int n;cin>>n;
+    long long n;cin>>n;
     cout<<n<<" ";
     while(n!=1){
        if(n%2==0){n=n/2;cout<<n<<" ";}
